pyproject.toml
[project]
name = "chart"
version = "0.1.0"
description = ""
authors = [
    {name = "zyuvszqk",email = "zumwachang479@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "socketify (>=0.0.31,<0.0.32)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
